Roster and functions
Fixed the roster and function label loading to WiThrottle.  Functions are now intercepted in the ESP.  F0-F4 are directed to the Arduino per the original DCC Base Station design.  Volume up/down and sound selections data are controlled by F5-F28 which are directed to the Base Station with the J prefix which calls subroutines in SerialCommand to send signals out on Serial1 to the DF Player Mini.  Sound selection now occurs correctly from WiThrottle on the iPhone to the ESP through the Arduino to the DF Player Mini.  The Arduino mega communicates to the ESP-01 on Serial3 in order to free up Serial for monitoring and debugging.  Use of JMRI via USB requires changing the COMM_INTERFACE constant from 4 back to 0.  This can be done in Config.h under DEFINE COMMUNICATIONS INTERFACE.